WIP: Generic Send Offload (GSO) Support
This change implements a minimal version of GSO support [1][2][3]
for Linux, which improves the effiency of sending data.

The approach taken in this change is to create a buffer which contains
multiple datagrams in `Connection::poll_transmit`. This was picked
over trying to merge packets in the endpoint task, since packets
seem to need to be padded to a common segment size in order to
make them sendable via GSO.

The change tries to minimize the impact on existing code paths,
and still tries to keep packet creation is independent as possible.
GSO is currently only used for data spaces, due to a less
understood impact on other packet spaces and packet types.
It can be extended for those.

There are a few tricky things regarding padding code - which
partially assumed StartOfPacket == StartOfDataram == StartOfSendBuffer,
which is not true anymore. This likely requires additional review.

In addition to this the mechanism on whether/how to configure GSO
should be discussed.

The change also isn't perfect in a sense that the buffer for the
outgoing datagram needs to be resized during sending - potentially even
multiple times - which causes extra unnecessary copies and degrades
performance with a high number of segments. That could potentially be
mitigated by using a suitable buffer pool.

**No GSO:**

```
Sent 1073741824 bytes on 1 streams in 4.31s (237.66 MiB/s)
```

**With GSO (up to 8 packets):**

```
Sent 1073741824 bytes on 1 streams in 3.02s (339.18 MiB/s)
```

[1] http://vger.kernel.org/lpc_net2018_talks/willemdebruijn-lpc2018-udpgso-paper-DRAFT-1.pdf
[2] http://vger.kernel.org/lpc_net2018_talks/willemdebruijn-lpc2018-udpgso-presentation-20181104.pdf
[3] https://lwn.net/Articles/752956/